[CI/Build]: Replace linting and formatting management with pre-commit (#691)
* Move linting to pre-commit

Using pre-commit for linting, formatting and type checking.

This replaces the format.sh bash script and tools dependencies
in requirements-lint file to be managed by pre-commit ecosystem.
It also provides checking at commit time.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Replace existing linting and formattion GitHub workflow

Linting, formatting and static checking is now 1 workflow using
pre-commit to run the relevant tools.

It replaces the existing workflows for mypy, ruff and yapf.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Enable yapf formatter

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add ruff for linting and formatting

This commit also removes yapf for formatting. as there is no need for
2 different formatters.

Ruff formatter is an extremely fast Python formatter, written in Rust.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add updates after rebase

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add Python import format handling with isort

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Renable ruff linting and formatting

Disabled ruff when adding isort as it was clashing. It is now
working after the isort changes were made.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add codespell

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add clang-format

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Add MyPy static checker

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Update documentation

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Remove redundant workflows

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Updates after rebase

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

---------

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>