[Test]: Add integration testing for LMCache integration with vLLM (#930)
* Add functional test for lmcache/vllm-openai container image

Add functional test in the CI to avoid regressions between LMCache and vLLM.

This script will be the basis for all functional tests.
It will start with one test that prompts a model on the lmcache/vllm-openai API server.

More tests can be added on in subsequent PRs.

Closes #801

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* test pipeline

Signed-off-by: Shaoting Feng <shaotingf@uchicago.edu>

* Add pipeline yml

Signed-off-by: Shaoting Feng <shaotingf@uchicago.edu>

* Add buildkit support

Signed-off-by: Shaoting Feng <shaotingf@uchicago.edu>

* Troubleshoot the script in pipeline

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Test hf_token

Signed-off-by: Shaoting Feng <shaotingf@uchicago.edu>

* Add HF access token

Access token required for downloading model from HuggingFace.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Update test script to take server wait timeout as a flag

Need to be able to pass the timeout as OpenAI server takes longer
to start of buildkite host

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Tidyup up script

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

* Modifying from functional to integration tests

The tests are testing the integration of LMCache with vLLM
and therefore are better defined as such rather than functional tests.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>

---------

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>
Signed-off-by: Shaoting Feng <shaotingf@uchicago.edu>
Co-authored-by: Shaoting Feng <shaotingf@uchicago.edu>